Make IPC API the primary kicad parser, demote SWIG to kicad_swig
- Move legacy SWIG parser (PcbnewParser + the ActionPlugin) to
  ecad/kicad_swig.py and promote the IPC API parser into ecad/kicad.py
  so the IPC backend is the default going forward.
- generate_interactive_bom.py: route .kicad_pcb through the IPC parser
  when an API server context is present (KICAD_API_SOCKET), otherwise
  fall back to the SWIG parser which loads the board file from disk.
- Skip registering the legacy SWIG ActionPlugin on KiCad 9+ (where the
  IPC plugin manifest provides the iBOM action) to avoid a duplicate
  toolbar button.

I hate work, and may the Claude be with you.
