+++ InteractiveHtmlBom/ecad/__init__.py
@@ -4,7 +4,19 @@ import os
 def get_parser_by_extension(file_name, config, logger):
     ext = os.path.splitext(file_name)[1]
     if ext == '.kicad_pcb':
-        return get_kicad_parser(file_name, config, logger)
+        # Prefer the IPC API parser, but only when an API server context is
+        # actually present (KICAD_API_SOCKET is set by a running KiCad or by
+        # kicad-cli's server mode, and kipy picks it up automatically).
+        # Without it we must not connect to some unrelated running KiCad and
+        # parse its open board; we fall back to the legacy SWIG parser, which
+        # is the only one that loads the board file straight from disk.
+        if os.environ.get('KICAD_API_SOCKET'):
+            try:
+                return get_kicad_parser(file_name, config, logger)
+            except Exception as e:
+                logger.info("IPC API unavailable (%s), "
+                            "falling back to SWIG parser." % e)
+        return get_kicad_swig_parser(file_name, config, logger)
     elif ext == '.json':
         """.json file may be from EasyEDA or a generic json format"""
         import io
@@ -21,14 +33,18 @@ def get_parser_by_extension(file_name, config, logger):
         return None
 
 
-def get_kicad_parser(file_name, config, logger, board=None):
-    from .kicad import PcbnewParser
-    return PcbnewParser(file_name, config, logger, board)
-
-
-def get_kicad_ipc_parser(file_name, config, logger, kicad=None, board=None):
-    from .kicad_ipc import IpcApiParser
+def get_kicad_parser(file_name, config, logger, kicad=None, board=None):
+    from .kicad import IpcApiParser
     return IpcApiParser(file_name, config, logger, kicad, board)
+
+
+# Backwards compatible alias for the IPC parser.
+get_kicad_ipc_parser = get_kicad_parser
+
+
+def get_kicad_swig_parser(file_name, config, logger, board=None):
+    from .kicad_swig import PcbnewParser
+    return PcbnewParser(file_name, config, logger, board)
 
 
 def get_easyeda_parser(file_name, config, logger):
